Big UI redesign
epigram now looks more white.

Most red and white colours of the app have been inverted to give a more clean and professional look and feel to the app, the app title is now centred with a dummy burger ready for implementation
the home screen app bar now partially collapses if you scroll on the app bar itself, it does not work if you scroll on the recycler view.
constant padding style and font changes